+++ packages/frontend/src/components/FooterComponent.vue
@@ -1,0 +1,6 @@
+<template>
+  <footer class="bg-text text-text-inverse block min-h-12 p-4">
+    Created by Patrik Prinz<br />
+    patrik.prinz@protonmail.com
+  </footer>
+</template>