+++ packages/frontend/src/modules/admin/services/admin.service.ts
@@ -2,7 +2,7 @@ import type { Axios } from 'axios';
 import {
   getUsersResponseSchema,
   getGroupsResponseSchema,
-  createGroupResponseSchema,
+  groupResponseSchema,
   getRolesResponseSchema,
 } from './adminService.schema';
 import { HttpError } from '@/shared/httpClient/http.errors';
@@ -49,7 +49,7 @@ export class AdminService {
       toast.error('Log in as admin to view this section.');
       return undefined;
     }
-    const validatedData = createGroupResponseSchema.safeParse(response.data);
+    const validatedData = groupResponseSchema.safeParse(response.data);
     if (validatedData.success) {
       return validatedData.data as Group;
     }
@@ -115,11 +115,13 @@ export class AdminService {
   public async assignRole(
     roleId: string,
     userId: string,
+    groupId?: string,
   ): Promise<string | undefined> {
     const response = await this.client.post('admin/roles/assign', {
       userRole: {
         userId: userId,
         roleId: roleId,
+        groupId: groupId,
       },
     });
     if (response.status == 401 || response.status == 403) {
@@ -129,11 +131,16 @@ export class AdminService {
     return response.data as string;
   }
 
-  public async unassignRole(roleId: string, userId: string): Promise<boolean> {
+  public async unassignRole(
+    roleId: string,
+    userId: string,
+    groupId?: string,
+  ): Promise<boolean> {
     const response = await this.client.post('admin/roles/unassign', {
       userRole: {
         userId: userId,
         roleId: roleId,
+        groupId: groupId,
       },
     });
     if (response.status == 401 || response.status == 403) {