+++ packages/frontend/src/modules/admin/services/adminService.schema.ts
@@ -8,17 +8,12 @@ export const getUsersResponseSchema = z.array(
   }),
 );
 
-export const getGroupsResponseSchema = z.array(
-  z.object({
-    id: z.string(),
-    name: z.string(),
-  }),
-);
-
-export const createGroupResponseSchema = z.object({
+export const groupResponseSchema = z.object({
   id: z.string(),
   name: z.string(),
 });
+
+export const getGroupsResponseSchema = z.array(groupResponseSchema);
 
 export const getRolesResponseSchema = z.array(
   z.object({