+++ packages/server/src/modules/auth/services/roles.service.ts
@@ -2,6 +2,7 @@ import { inject, injectable } from 'tsyringe';
 import { PermissionError } from '../../../shared/error_handler/errors.js';
 import type {
   Role,
+  RoleCode,
   User,
   UserRole,
 } from '../../../shared/types/auth/auth.types.js';
@@ -25,12 +26,13 @@ export class RolesService {
    */
   public async hasRole(
     user: User,
-    roles: string[] = ['admin'],
+    roles: RoleCode[] = ['admin'],
+    groupId: string = '0',
   ): Promise<boolean> {
     if (!roles.find((e) => e == 'admin')) {
       roles.push('admin');
     }
-    return await this.repository.hasRole(user.id, roles);
+    return await this.repository.hasRole(user.id, roles, groupId);
   }
 
   public async listRoles(author: User) {
@@ -46,6 +48,7 @@ export class RolesService {
       const result = await this.repository.assignRole(
         userRole.userId,
         userRole.roleId,
+        userRole.groupId,
       );
       return result;
     }